+++ changelog.md
@@ -9,6 +9,7 @@
 ### Changes
 
 - [#2957](https://github.com/ignite/cli/pull/2957) Change generate commands to print the path to the generated code.
+- [#2981](https://github.com/ignite/cli/issues/2981) Change CLI to also search chain binary in Go binary path.
 
 ## [`v0.25.1`](https://github.com/ignite/cli/releases/tag/v0.25.1)